requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5 